Refacror and remove unused files

+++ gen/gen.mjs
@@ -1,8 +1,9 @@
 import path from 'path'
-import {rmdir} from './utils/rmdir.mjs'
-import {cpdir} from './utils/cpdir.mjs'
 import fs from 'fs'
 import parse, {Globals, Native, Type, Variable} from 'jass-to-ast'
+
+import {rmdir} from './utils/rmdir.mjs'
+import {cpdir} from './utils/cpdir.mjs'
 import nativeAdd, {
     TriggerActions, TriggerActionsMap, TriggerActionStrings,
     TriggerCalls, TriggerCallsMap, TriggerCallStrings,
@@ -16,54 +17,68 @@ import typeAdd, {TriggerTypes, TriggerTypesMap} from './utils/type-add.mjs'
 import capitalize from './utils/capitalize.mjs'
 import {paramAdd, TriggerParams} from './utils/param.mjs'
 
+let commonApi = process.argv[2]
+if (!commonApi) {
+    console.error('Usage: npm run generate ./data/common.vj')
+    process.exit(1)
+}
+commonApi = path.resolve(commonApi)
+if (!fs.existsSync(commonApi)) {
+    console.error(`File is not found: ${commonApi}`)
+    process.exit(1)
+}
+
 const UI = path.join('.', 'dist', 'UI')
-
-rmdir(UI)
-cpdir(path.join('.', 'data', 'umswe', 'UI'), UI)
-const read = p => fs.readFileSync(p, {encoding: 'utf8', flag: 'r'}).replace(/\r\n/g, '\n')
-
-
 const TriggerData = path.join(UI, 'TriggerData.txt')
 const TriggerStrings = path.join(UI, 'TriggerStrings.txt')
 
-let cat = ''
-for (const s of fs.readFileSync(TriggerData, {encoding: 'utf8', flag: 'r'}).split(/\r\n|\n/)) {
-    const cm = s.match(/^\[([a-z0-9]+)\s*]/i)
-    if (cm) cat = cm[1].trim()
+rmdir(UI)
+cpdir(path.join('.', 'data', 'umswe', 'UI'), UI)
 
-    const m = s.match(/^([a-z][a-z0-9_]+)=/i)
-    if (!m) continue
-    const name = m[1]
+const read = p => fs.readFileSync(p, {encoding: 'utf8', flag: 'r'}).replace(/\r\n/g, '\n')
 
-    if (cat === 'TriggerActions') TriggerActionsMap[name] = true
-    if (cat === 'TriggerCalls') TriggerCallsMap[name] = true
-    if (cat === 'TriggerTypes') TriggerTypesMap[name] = true
-    if (cat === 'TriggerEvents') TriggerEventsMap[name] = true
+/**
+ * @param {string} file
+ * @param {string[]} list
+ * @param {string} section
+ */
+const sortAppend = (file, list, section) => {
+    list.sort((a, b) => a.localeCompare(b))
+    fileDataAppend(file, list.join('\n'), section)
 }
 
 /**
- * @param {number} part
- * @param {Native} native
+ * @param {string[]} sections
+ * @param {import('jass-to-ast').Native} native
  */
-/** @param {string[]} sections @param {import('jass-to-ast').Native} native */
 const nativeAddWrapper = (sections, native) => {
-	for (const sec of sections) nativeAdd(native, sec)
+    for (const section of sections) nativeAdd(native, section)
 }
 
-const i2a = i => {
-    switch (i) {
-        case 0 :
-            return 'unit'
-        case 1:
-            return 'item'
-        case 2:
-            return 'destructable'
+const WIDGET_TYPES = ['unit', 'item', 'destructable']
+const i2a = i => WIDGET_TYPES[i]
+
+{
+    let cat = ''
+    for (const s of read(TriggerData).split('\n')) {
+        const cm = s.match(/^\[([a-z0-9]+)\s*]/i)
+        if (cm) cat = cm[1].trim()
+
+        const m = s.match(/^([a-z][a-z0-9_]+)=/i)
+        if (!m) continue
+        const name = m[1]
+
+        if (cat === 'TriggerActions') TriggerActionsMap[name] = true
+        if (cat === 'TriggerCalls') TriggerCallsMap[name] = true
+        if (cat === 'TriggerTypes') TriggerTypesMap[name] = true
+        if (cat === 'TriggerEvents') TriggerEventsMap[name] = true
     }
 }
 
-for (const node of parse(read(path.join('.', 'data', 'common.j')))) {
+for (const node of parse(read(commonApi))) {
     if (node instanceof Native) {
         const sections = nativeSection(node)
+
         if (node.params) {
             const list = []
             for (const p of node.params) {
@@ -113,24 +128,12 @@ for (const node of parse(read(path.join('.', 'data', 'common.j')))) {
     }
 }
 
-
-TriggerActions.sort((a, b) => a.localeCompare(b))
-fileDataAppend(TriggerData, TriggerActions.join('\n'), 'TriggerActions')
-
-TriggerCalls.sort((a, b) => a.localeCompare(b))
-fileDataAppend(TriggerData, TriggerCalls.join('\n'), 'TriggerCalls')
-
-TriggerTypes.sort((a, b) => a.localeCompare(b))
-fileDataAppend(TriggerData, TriggerTypes.join('\n'), 'TriggerTypes')
-
-TriggerCategories.sort((a, b) => a.localeCompare(b))
-fileDataAppend(TriggerData, TriggerCategories.join('\n'), 'TriggerCategories')
-
-TriggerParams.sort((a, b) => a.localeCompare(b))
-fileDataAppend(TriggerData, TriggerParams.join('\n'), 'TriggerParams')
-
-TriggerEvents.sort((a, b) => a.localeCompare(b))
-fileDataAppend(TriggerData, TriggerEvents.join('\n'), 'TriggerEvents')
+sortAppend(TriggerData, TriggerActions, 'TriggerActions')
+sortAppend(TriggerData, TriggerCalls, 'TriggerCalls')
+sortAppend(TriggerData, TriggerTypes, 'TriggerTypes')
+sortAppend(TriggerData, TriggerCategories, 'TriggerCategories')
+sortAppend(TriggerData, TriggerParams, 'TriggerParams')
+sortAppend(TriggerData, TriggerEvents, 'TriggerEvents')
 
 fileDataAppend(TriggerStrings, TriggerActionStrings.join('\n'), 'TriggerActionStrings')
 fileDataAppend(TriggerStrings, TriggerCallStrings.join('\n'), 'TriggerCallStrings')
